prepare exercise implementation; refactoring

+++ src/script.js
@@ -3,9 +3,10 @@ import {
   ADDED_SESSION_CARD_LABEL,
   DELETE_SESSION_BUTTON_LABEL,
   ADD_BUTTON_LABEL,
-  SESSIONS
+  SESSIONS,
 } from './constants.js';
 import { weekdaysData } from './data.js';
+import { renderExercisePage } from './exercise.js';
 
 const myStorage = window.localStorage;
 
@@ -17,28 +18,31 @@ const weekdays = JSON.parse(localStorage.getItem('weekdays'));
 
 function init() {
   function addDeleteSessionCardLogic(deleteButton) {
-    //find surrounding div of delete button in order to remove it from the DOM
-    const deletableCard = deleteButton.parentNode.parentNode;
-    const weekday = deletableCard.parentNode;
-    deletableCard.remove();
+    const shallDelete = confirm('Bist du sicher?');
+    if (shallDelete) {
+      //find surrounding div of delete button in order to remove it from the DOM
+      const deletableCard = deleteButton.parentNode.parentNode;
+      const weekday = deletableCard.parentNode;
+      deletableCard.remove();
 
-    //find index of affected weekday in weekday array, manipulate the session string to ''
-    const weekdayID = weekday.id;
-    weekdays.map((weekday) => {
-      if (weekday.day === weekdayID) {
-        weekday.session = '';
-      }
-    });
+      //find index of affected weekday in weekday array, manipulate the session string to ''
+      const weekdayID = weekday.id;
+      weekdays.map((weekday) => {
+        if (weekday.day === weekdayID) {
+          weekday.session = '';
+        }
+      });
 
-    // delete session card from local storage
-    myStorage.setItem('weekdays', JSON.stringify(weekdays));
+      // delete session card from local storage
+      myStorage.setItem('weekdays', JSON.stringify(weekdays));
 
-    //re-create the add session button and make sure it gets an eventlistener assignes
-    const addButton = document.createElement('button');
-    addButton.innerText = ADD_SESSION_BUTTON_LABEL;
-    addButton.classList.add('add-session');
-    weekday.appendChild(addButton);
-    addAddSessionCardLogic(addButton);
+      //re-create the add session button and make sure it gets an eventlistener assignes
+      const addButton = document.createElement('button');
+      addButton.innerText = ADD_SESSION_BUTTON_LABEL;
+      addButton.classList.add('add-session');
+      weekday.appendChild(addButton);
+      addAddSessionCardLogic(addButton);
+    }
   }
 
   let clickedWeekdayID = null;
@@ -70,8 +74,8 @@ function init() {
     if (session) {
       const sessionTitle = SESSIONS[session].title;
       const sessionCard = ` 
-        <div class="added-card">
-          <div class="added-card-color" id="${session}-card"></div>
+        <div class="added-card" id="${session}-card">
+          <div class="added-card-color"></div>
           <div class="added-card-subcontainer">
             <div class="added-card-headline">${sessionTitle}</div>
             <div class="added-card-subtext">
@@ -83,10 +87,19 @@ function init() {
       `;
       const weekday = document.getElementById(day);
       weekday.insertAdjacentHTML('beforeend', sessionCard);
+
+      // add session card event listener
+      const sessionCardElement = weekday.querySelector('.added-card');
+      sessionCardElement.addEventListener('click', () => {
+        renderExercisePage();
+      });
+
+      // add delete card event listener
       const deleteButton = weekday.getElementsByClassName('delete')[0];
-      deleteButton.addEventListener('click', () =>
-        addDeleteSessionCardLogic(deleteButton)
-      );
+      deleteButton.addEventListener('click', (event) => {
+        event.stopPropagation();
+        addDeleteSessionCardLogic(deleteButton);
+      });
     } else {
       const addSessionButton = `<button class="add-session">${ADD_SESSION_BUTTON_LABEL}</button>`;
       const weekday = document.getElementById(`${day}`);
